Adding the coding and result files
Speech_detection_audio_classifier.ipynb is the python notebook that includes all of the codes for the audio classification model using Google's audio dataset. The Speech_Detecting_Audio_Classifier.pth is the saved model stored after training the model. And the rest .png files are the results in visual form.